app/requirements.txt
WTForms<=3.0.1
Flask_Login<=0.5.0
SQLAlchemy<=1.4.31
Flask<=1.1.2
Jinja2<=3.0.3
phonenumbers<=8.12.43
Flask_Security_Too==3.3.0rc3
Flask_WTF<=1.0.0
Flask_SQLAlchemy<=2.5.1
flask_avatars<=0.2.2
requests<=2.27.1
flask-jsontools<=0.1.7
flask_script<=2.0.6
Flask-Migrate<=3.1.0
flask-cors<=3.0.10
jsonschema[format]<=4.4.0
redis<=4.1.3
flask-caching<=1.10.1
dogpile.cache<=1.1.5
bcrypt<=3.2.0
werkzeug==0.16.1
cachetools<=5.0.0
email_validator<=1.1.3
waitress<=2.0.0
itsdangerous==2.0.1
gunicorn
